@@ -2,41 +2,116 @@ package com.baeldung.jersey.server.rest;
 
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.containsString;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
 
+import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.Application;
+import javax.ws.rs.core.Form;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 import org.glassfish.jersey.test.JerseyTest;
-import org.junit.Assert;
+import org.glassfish.jersey.test.TestProperties;
 import org.junit.Test;
 
 import com.baeldung.jersey.server.config.ViewApplicationConfig;
+import com.baeldung.jersey.server.model.Fruit;
+import com.baeldung.jersey.server.providers.FruitExceptionMapper;
 
 public class FruitResourceIntegrationTest extends JerseyTest {
 
     @Override
     protected Application configure() {
-        return new ViewApplicationConfig();
+        enable(TestProperties.LOG_TRAFFIC);
+        enable(TestProperties.DUMP_ENTITY);
+
+        ViewApplicationConfig config = new ViewApplicationConfig();
+        config.register(FruitExceptionMapper.class);
+        return config;
     }
 
     @Test
-    public void testAllFruit() {
+    public void givenGetAllFruit_whenCorrectRequest_thenAllTemplateInvoked() {
         final String response = target("/fruit/all").request()
             .get(String.class);
-        Assert.assertThat(response, allOf(containsString("banana"), containsString("apple"), containsString("kiwi")));
+        assertThat(response, allOf(containsString("banana"), containsString("apple"), containsString("kiwi")));
     }
 
     @Test
-    public void testIndex() {
+    public void givenGetFruit_whenCorrectRequest_thenIndexTemplateInvoked() {
         final String response = target("/fruit").request()
             .get(String.class);
-        Assert.assertThat(response, containsString("Welcome Fruit Index Page!"));
+        assertThat(response, containsString("Welcome Fruit Index Page!"));
     }
 
     @Test
-    public void testErrorTemplate() {
+    public void givenGetFruitByName_whenFruitUnknown_thenErrorTemplateInvoked() {
         final String response = target("/fruit/orange").request()
             .get(String.class);
-        Assert.assertThat(response, containsString("Error -  Fruit not found: orange!"));
+        assertThat(response, containsString("Error -  Fruit not found: orange!"));
+    }
+
+    @Test
+    public void givenCreateFruit_whenFormContainsNullParam_thenResponseCodeIsBadRequest() {
+        Form form = new Form();
+        form.param("name", "apple");
+        form.param("colour", null);
+        Response response = target("fruit/create").request(MediaType.APPLICATION_FORM_URLENCODED)
+            .post(Entity.form(form));
+
+        assertEquals("Http Response should be 400 ", 400, response.getStatus());
+        assertThat(response.readEntity(String.class), containsString("Fruit colour must not be null"));
+    }
+
+    @Test
+    public void givenUpdateFruit_whenFormContainsBadSerialParam_thenResponseCodeIsBadRequest() {
+        Form form = new Form();
+        form.param("serial", "2345-2345");
+
+        Response response = target("fruit/update").request(MediaType.APPLICATION_FORM_URLENCODED)
+            .put(Entity.form(form));
+
+        assertEquals("Http Response should be 400 ", 400, response.getStatus());
+        assertThat(response.readEntity(String.class), containsString("Fruit serial number is not valid"));
+    }
+
+    @Test
+    public void givenCreateFruit_whenFruitIsInvalid_thenResponseCodeIsBadRequest() {
+        Fruit fruit = new Fruit("Blueberry", "purple");
+        fruit.setWeight(1);
+
+        Response response = target("fruit/create").request(MediaType.APPLICATION_JSON_TYPE)
+            .post(Entity.entity(fruit, MediaType.APPLICATION_JSON_TYPE));
+
+        assertEquals("Http Response should be 400 ", 400, response.getStatus());
+        assertThat(response.readEntity(String.class), containsString("Fruit weight must be 10 or greater"));
+    }
+
+    @Test
+    public void givenFruitExists_whenSearching_thenResponseContainsFruit() {
+        Fruit fruit = new Fruit();
+        fruit.setName("strawberry");
+        fruit.setWeight(20);
+        Response response = target("fruit/create").request(MediaType.APPLICATION_JSON_TYPE)
+            .post(Entity.entity(fruit, MediaType.APPLICATION_JSON_TYPE));
+
+        assertEquals("Http Response should be 204 ", 204, response.getStatus());
+
+        final String json = target("fruit/search/strawberry").request()
+            .get(String.class);
+        assertThat(json, containsString("{\"name\":\"strawberry\",\"weight\":20}"));
+    }
+
+    @Test
+    public void givenFruit_whenFruitIsInvalid_thenReponseContainsCustomExceptions() {
+        final Response response = target("fruit/exception").request()
+            .get();
+
+        assertEquals("Http Response should be 400 ", 400, response.getStatus());
+        String responseString = response.readEntity(String.class);
+        assertThat(responseString, containsString("exception.<return value>.colour size must be between 5 and 200"));
+        assertThat(responseString, containsString("exception.<return value>.name size must be between 5 and 200"));
     }
 
 }
