@@ -1,5 +1,6 @@
 package com.baeldung.spring.data.neo4j.config;
 
+import org.neo4j.ogm.config.Configuration.Builder;
 import org.neo4j.ogm.session.SessionFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -8,15 +9,14 @@ import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 
 @ComponentScan(basePackages = { "com.baeldung.spring.data.neo4j.services" })
 @Configuration
-@EnableNeo4jRepositories(basePackages = "com.baeldung.spring.data.neo4j.repostory")
+@EnableNeo4jRepositories(basePackages = "com.baeldung.spring.data.neo4j.repository")
 public class MovieDatabaseNeo4jConfiguration {
 
     public static final String URL = System.getenv("NEO4J_URL") != null ? System.getenv("NEO4J_URL") : "http://neo4j:movies@localhost:7474";
 
     @Bean
     public org.neo4j.ogm.config.Configuration getConfiguration() {
-        org.neo4j.ogm.config.Configuration config = new org.neo4j.ogm.config.Configuration();
-        config.driverConfiguration().setDriverClassName("org.neo4j.ogm.drivers.http.driver.HttpDriver").setURI(URL);
+    	org.neo4j.ogm.config.Configuration config = new Builder().uri(URL).build();
         return config;
     }
 
