@@ -1,12 +1,13 @@
 package com.baeldung.spring.data.neo4j.domain;
 
-import org.neo4j.ogm.annotation.GraphId;
+import org.neo4j.ogm.annotation.GeneratedValue;
+import org.neo4j.ogm.annotation.Id;
 import org.neo4j.ogm.annotation.NodeEntity;
 import org.neo4j.ogm.annotation.Relationship;
 
 @NodeEntity
 public class Car {
-    @GraphId
+    @Id @GeneratedValue
     private Long id;
 
     private String make;
