@@ -3,7 +3,8 @@ package com.baeldung.spring.data.neo4j.domain;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.voodoodyne.jackson.jsog.JSOGGenerator;
 import org.neo4j.ogm.annotation.EndNode;
-import org.neo4j.ogm.annotation.GraphId;
+import org.neo4j.ogm.annotation.GeneratedValue;
+import org.neo4j.ogm.annotation.Id;
 import org.neo4j.ogm.annotation.RelationshipEntity;
 import org.neo4j.ogm.annotation.StartNode;
 
@@ -12,8 +13,8 @@ import java.util.Collection;
 @JsonIdentityInfo(generator = JSOGGenerator.class)
 @RelationshipEntity(type = "ACTED_IN")
 public class Role {
-    @GraphId
-    Long id;
+    @Id @GeneratedValue
+	Long id;
     private Collection<String> roles;
     @StartNode
     private Person person;
