@@ -13,7 +13,7 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyEmitter
 
 @Controller
 public class ResponseBodyEmitterController {
-    private ExecutorService nonBlockingService = Executors.newSingleThreadExecutor();
+    private ExecutorService nonBlockingService = Executors.newCachedThreadPool();
 
     @GetMapping(Constants.API_RBE)
     public ResponseEntity<ResponseBodyEmitter> handleRbe() {
