@@ -27,19 +27,25 @@ public class ScopesConfig {
 
     @Bean
     @Scope(value = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
-    public HelloMessageGenerator requestMessage() {
+    public HelloMessageGenerator requestScopedBean() {
         return new HelloMessageGenerator();
     }
 
     @Bean
     @Scope(value = WebApplicationContext.SCOPE_SESSION, proxyMode = ScopedProxyMode.TARGET_CLASS)
-    public HelloMessageGenerator sessionMessage() {
+    public HelloMessageGenerator sessionScopedBean() {
         return new HelloMessageGenerator();
     }
 
     @Bean
     @Scope(value = WebApplicationContext.SCOPE_APPLICATION, proxyMode = ScopedProxyMode.TARGET_CLASS)
-    public HelloMessageGenerator globalSessionMessage() {
+    public HelloMessageGenerator applicationScopedBean() {
+        return new HelloMessageGenerator();
+    }
+
+    @Bean
+    @Scope(scopeName = "websocket", proxyMode = ScopedProxyMode.TARGET_CLASS)
+    public HelloMessageGenerator websocketScopedBean() {
         return new HelloMessageGenerator();
     }
 
