@@ -1,13 +1,15 @@
 package org.baeldung.properties;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.annotation.ComponentScan;
 
-@EnableAutoConfiguration
-@ComponentScan(basePackageClasses = ConfigProperties.class)
+@SpringBootApplication
+@ComponentScan(basePackageClasses = { ConfigProperties.class, JsonProperties.class, CustomJsonProperties.class })
 public class ConfigPropertiesDemoApplication {
     public static void main(String[] args) {
-        SpringApplication.run(ConfigPropertiesDemoApplication.class);
+        new SpringApplicationBuilder(ConfigPropertiesDemoApplication.class).initializers(new JsonPropertyContextInitializer())
+            .run();
     }
+
 }
