@@ -2,13 +2,16 @@ package com.baeldung;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+import com.baeldung.dao.repositories.impl.ExtendedRepositoryImpl;
 
 @SpringBootApplication
+@EnableJpaRepositories(repositoryBaseClass = ExtendedRepositoryImpl.class)
 public class Application {
-    private static ApplicationContext applicationContext;
 
     public static void main(String[] args) {
-        applicationContext = SpringApplication.run(Application.class, args);
+        SpringApplication.run(Application.class, args);
     }
+
 }
