@@ -1,37 +1,52 @@
 'use strict';
 
 angularApp
-    .controller('socketController', function ($scope, SocketService) {
+.controller('socketController', function ($scope, SocketService) {
 
-        $scope.stompClient = null;
-        $scope.sendEndpoint = '/secured/chat';
-        $scope.subscribeEndpoint = '/secured/history';
-        $scope.elems = {
-            connect: 'connect',
-            from: 'from',
-            text: 'text',
-            disconnect: 'disconnect',
-            conversationDiv: 'conversationDiv',
-            response: 'response'
-        };
+  /**
+   * URL mapping endpoints.
+   */
 
-        $scope.connect = function (context) {
-            $scope.sendEndpoint = '/secured/chat';
-            $scope.sendEndpoint = context + $scope.sendEndpoint ;
-            $scope.stompClient = SocketService.connect($scope.sendEndpoint, $scope.elems);
-        };
+  var SECURED_CHAT = '/secured/chat';
+  var SECURED_CHAT_HISTORY = '/secured/history';
+  var SECURED_CHAT_ROOM = '/secured/room';
+  var SECURED_CHAT_SPECIFIC_USER = '/secured/user/queue/specific-user';
 
-        $scope.subscribe = function () {
-            $scope.stompClient.subscribe($scope.subscribeEndpoint, function (msgOut) {
-                SocketService.messageOut(JSON.parse(msgOut.body), $scope.elems);
-            });
-        };
+  var opts = {
+    from: 'from',
+    to: 'to',
+    text: 'text',
+    disconnect: 'disconnect',
+    conversationDiv: 'conversationDiv',
+    response: 'response'
+  };
 
-        $scope.disconnect = function () {
-            SocketService.disconnect($scope.elems, $scope.stompClient);
-        };
+  $scope.sendEndpoint = '';
+  $scope.stompClient = null;
 
-        $scope.sendMessage = function () {
-            SocketService.sendMessage( $scope.elems, $scope.stompClient, $scope.sendEndpoint);
-        };
-    });
+  /**
+   * Broadcast to All Users.
+   */
+
+  $scope.connectAll = function (context) {
+    $scope.sendEndpoint = context + SECURED_CHAT;
+    $scope.stompClient = SocketService.connect($scope.sendEndpoint , opts, true);
+  };
+
+  $scope.subscribeAll = function () { SocketService.subscribeToAll($scope.stompClient, SECURED_CHAT_HISTORY, opts); };
+
+  /**
+   * Broadcast to Specific User.
+   */
+
+  $scope.connectSpecific = function (context) {
+    $scope.sendEndpoint = context + SECURED_CHAT_ROOM;
+    $scope.stompClient = SocketService.connect(context + SECURED_CHAT_ROOM, opts, false);
+  };
+
+  $scope.subscribeSpecific = function () { SocketService.subscribeToSpecific($scope.stompClient, SECURED_CHAT_SPECIFIC_USER, opts); };
+
+  $scope.disconnect = function () { SocketService.disconnect(opts, $scope.stompClient); };
+
+  $scope.sendMessage = function () { SocketService.sendMessage(opts, $scope.stompClient, $scope.sendEndpoint); };
+});
