@@ -1,13 +1,12 @@
 'use strict';
 
 angularApp
-    .controller('successController', function ($scope) {
+.controller('successController', function ($scope) {
+  $scope.successMsg = '';
 
-        $scope.successMsg = '';
+  $scope.initialize = function () {
+    $scope.successMsg = "You've logged in!";
+  };
 
-        $scope.initialize = function () {
-            $scope.successMsg = "You've logged in!";
-        };
-
-        $scope.initialize();
-    });
+  $scope.initialize();
+});
