@@ -1,19 +1,16 @@
 'use strict';
 
 angularApp
-	.config(function ($routeProvider) {
-		$routeProvider
-			.when('/index', {
-				controller: 'indexController'
-			})
-			.when('/login', {
-				controller: 'loginController'
-			})
-			.when('/sockets', {
-				controller: 'socketController'
-			})
-			.when('/success', {
-				controller: 'successController'
-			})
-			.otherwise('/index');
-	});
+.config(function ($routeProvider) {
+  $routeProvider
+  .when('/index', {
+    controller: 'indexController'
+  })
+  .when('/sockets', {
+    controller: 'socketController'
+  })
+  .when('/success', {
+    controller: 'successController'
+  })
+  .otherwise('/index');
+});
